Support additional attributes extractors
