+++ instrumentation/netty/netty-4-common/library/src/main/java/io/opentelemetry/instrumentation/netty/v4/common/internal/HttpSchemeUtil.java
@@ -6,6 +6,7 @@
 package io.opentelemetry.instrumentation.netty.v4.common.internal;
 
 import io.netty.channel.ChannelHandler;
+import io.opentelemetry.instrumentation.netty.v4.common.HttpRequestAndChannel;
 
 /**
  * This class is internal and is hence not for public use. Its APIs are unstable and can change at
