+++ instrumentation/netty/netty-4.1/library/src/main/java/io/opentelemetry/instrumentation/netty/v4_1/NettyClientTelemetryBuilder.java
@@ -6,11 +6,14 @@
 package io.opentelemetry.instrumentation.netty.v4_1;
 
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
+import io.netty.handler.codec.http.HttpResponse;
 import io.opentelemetry.api.OpenTelemetry;
+import io.opentelemetry.instrumentation.api.instrumenter.AttributesExtractor;
+import io.opentelemetry.instrumentation.netty.v4.common.HttpRequestAndChannel;
 import io.opentelemetry.instrumentation.netty.v4.common.internal.client.NettyClientInstrumenterFactory;
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 
 /** A builder of {@link NettyClientTelemetry}. */
 public final class NettyClientTelemetryBuilder {
@@ -18,7 +21,8 @@ public final class NettyClientTelemetryBuilder {
   private final OpenTelemetry openTelemetry;
   private List<String> capturedRequestHeaders = Collections.emptyList();
   private List<String> capturedResponseHeaders = Collections.emptyList();
-  private Map<String, String> peerServiceMapping = Collections.emptyMap();
+  private final List<AttributesExtractor<HttpRequestAndChannel, HttpResponse>>
+      additionalAttributesExtractors = new ArrayList<>();
 
   NettyClientTelemetryBuilder(OpenTelemetry openTelemetry) {
     this.openTelemetry = openTelemetry;
@@ -49,14 +53,13 @@ public final class NettyClientTelemetryBuilder {
   }
 
   /**
-   * Configures the mapping between peer names and peer services used when capturing peer service
-   * span attributes.
-   *
-   * @param peerServiceMapping A map of peer names to peer services.
+   * Adds an additional {@link AttributesExtractor} to invoke to set attributes to instrumented
+   * items.
    */
   @CanIgnoreReturnValue
-  public NettyClientTelemetryBuilder setPeerServiceMapping(Map<String, String> peerServiceMapping) {
-    this.peerServiceMapping = peerServiceMapping;
+  public NettyClientTelemetryBuilder addAttributesExtractor(
+      AttributesExtractor<HttpRequestAndChannel, HttpResponse> attributesExtractor) {
+    additionalAttributesExtractors.add(attributesExtractor);
     return this;
   }
 
@@ -64,12 +67,8 @@ public final class NettyClientTelemetryBuilder {
   public NettyClientTelemetry build() {
     return new NettyClientTelemetry(
         new NettyClientInstrumenterFactory(
-            openTelemetry,
-            "io.opentelemetry.netty-4.1",
-            false,
-            false,
-            capturedRequestHeaders,
-            capturedResponseHeaders,
-            peerServiceMapping));
+                openTelemetry, "io.opentelemetry.netty-4.1", false, false, Collections.emptyMap())
+            .createHttpInstrumenter(
+                capturedRequestHeaders, capturedResponseHeaders, additionalAttributesExtractors));
   }
 }
