+++ instrumentation/reactor/reactor-netty/reactor-netty-1.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/reactornetty/v1_0/ReactorNettySingletons.java
@@ -76,8 +76,6 @@ public final class ReactorNettySingletons {
             INSTRUMENTATION_NAME,
             connectionTelemetryEnabled,
             false,
-            CommonConfig.get().getClientRequestHeaders(),
-            CommonConfig.get().getClientResponseHeaders(),
             CommonConfig.get().getPeerServiceMapping());
     CONNECTION_INSTRUMENTER = instrumenterFactory.createConnectionInstrumenter();
   }
